The experience
This project is an experience, inspired for the site http://cmd.fm, in
which step my afternoons and nights of work listening random
nastinesses.

I am a great fan of the site, and a great one loving of the terminal.
Being thus, I made my version in php, the language that not alive
without…

PHPlayer uses Joomla-Framework components.
Why not sf? Because not.

xxx